simulator/pyproject.toml
[tool.poetry]
name = "simulator"
version = "0.1.0"
description = "IoT MQTT device simulator"
authors = ["Seungbae Ji <seungbae.ji@gmail.com>"]
packages = [
    { include = "simulator", from = "src" },
]
include = [
    "config/",
]

[tool.poetry.scripts]
simulator = "simulator.main:main"

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = ">=0.115.13,<0.116.0"
paho-mqtt = ">=2.1.0,<3.0.0"
opentelemetry-sdk = ">=1.34.1,<2.0.0"
opentelemetry-exporter-otlp = ">=1.34.1,<2.0.0"
prometheus-client = ">=0.22.1,<0.23.0"
uvicorn = {version = "^0.29.0", extras = ["standard"]}
pydantic-settings = "^2.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"
ruff = "^0.12.0"
mypy = "^1.16.1"

[tool.pytest.ini_options]
addopts = "-ra -v"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
extend-exclude = ["tests/data/"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
strict = true
mypy_path = "src"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
